Server copies info from client side config
Summary:
- Fixed a issue with opening a file for writing, and not closing it to finish
- Unit Pixel Ratio under the homography header is copied over

Test:
By starting client and server app.py, and going through TrafficGUI to the
"Run" button, I was able to see in the TrafficCloud/project_dir/{UUID}/{UUID}.cfg
a section for homography, as well as the unit pixel ratio defined.  This value
should be the same as the one in the project_name.cfg file of the same directory.